Initial work on customizable colors
Need to create the JSON file to load these values from, hook up code to load the aforementioned JSON file on resource reload.
Maybe version the file (unless the values are fetched on a basis of a default value if the field doesn't exist)? Might version for the sake of consistency.
Maybe the colors for the boss death animation flimflam should be changeable as well?